Fixed display error cause by mismatch type: double -> unsigned long.

+++ src/compact_disk_monitor.cpp
@@ -47,12 +47,12 @@ CompactDiskMonitor::CompactDiskMonitor(QWidget *parent) : QWidget(parent)
 
     pointsNumber = int(statusBarMaxWidth / 5.4);
 
-    readSpeeds = new QList<double>();
+    readSpeeds = new QList<unsigned long>();
     for (int i = 0; i < pointsNumber; i++) {
         readSpeeds->append(0);
     }
 
-    writeSpeeds = new QList<double>();
+    writeSpeeds = new QList<unsigned long>();
     for (int i = 0; i < pointsNumber; i++) {
         writeSpeeds->append(0);
     }
@@ -80,7 +80,7 @@ void CompactDiskMonitor::changeTheme(QString )
     initTheme();
 }
 
-void CompactDiskMonitor::updateStatus(float tReadKbs, float tWriteKbs)
+void CompactDiskMonitor::updateStatus(unsigned long tReadKbs, unsigned long tWriteKbs)
 {
     totalReadKbs = tReadKbs;
     totalWriteKbs = tWriteKbs;
